requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cloudinary==1.23.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.3
Django==3.1.2
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-dropbox-storage==0.4.1
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.10.1
dropbox==10.10.0
Faker==4.14.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
